CONTRIB-8400: Inclusion of lib.php has side effects
* Move all non standard callback out of lib.php and in another namespace
* Update unit tests
* Removed deprecated lib callback
* Fixed a couple of coding issues (coding convention)
