+++ classes/locallib/config.php
@@ -25,6 +25,8 @@
 
 namespace mod_bigbluebuttonbn\locallib;
 
+use mod_bigbluebuttonbn\local\bbb_constants;
+
 defined('MOODLE_INTERNAL') || die();
 
 require_once($CFG->dirroot . '/mod/bigbluebuttonbn/locallib.php');
@@ -55,8 +57,8 @@ class config {
      */
     public static function defaultvalues() {
         return array(
-            'server_url' => (string) BIGBLUEBUTTONBN_DEFAULT_SERVER_URL,
-            'shared_secret' => (string) BIGBLUEBUTTONBN_DEFAULT_SHARED_SECRET,
+            'server_url' => (string) bbb_constants::BIGBLUEBUTTONBN_DEFAULT_SERVER_URL,
+            'shared_secret' => (string) bbb_constants::BIGBLUEBUTTONBN_DEFAULT_SHARED_SECRET,
             'voicebridge_editable' => false,
             'importrecordings_enabled' => false,
             'importrecordings_from_deleted_enabled' => false,
