+++ lib.php
@@ -23,55 +23,16 @@
  * @author    Jesus Federico  (jesus [at] blindsidenetworks [dt] com)
  * @author    Fred Dixon  (ffdixon [at] blindsidenetworks [dt] com)
  */
-
 defined('MOODLE_INTERNAL') || die;
+
+use mod_bigbluebuttonbn\local\bbb_constants;
+use mod_bigbluebuttonbn\local\helpers\files;
+use mod_bigbluebuttonbn\local\helpers\instance;
+use mod_bigbluebuttonbn\local\helpers\logs;
+use mod_bigbluebuttonbn\local\helpers\reset;
 
 global $CFG;
 
-// Do not declare new $CFG variables if unit tests are running
-// as it can cause "unexpected new $CFG->xxx value" warnings.
-if (!defined('PHPUNIT_TEST') || !PHPUNIT_TEST) {
-    if (!isset($CFG->bigbluebuttonbn)) {
-        $CFG->bigbluebuttonbn = array();
-    }
-
-    if (file_exists(dirname(__FILE__).'/config.php')) {
-        require_once(dirname(__FILE__).'/config.php');
-    }
-
-    /*
-     * DURATIONCOMPENSATION: Feature removed by configuration
-     */
-    $CFG->bigbluebuttonbn['scheduled_duration_enabled'] = 0;
-    /*
-     * Remove this block when restored
-     */
-}
-
-/** @var BIGBLUEBUTTONBN_DEFAULT_SERVER_URL string of default bigbluebutton server url */
-const BIGBLUEBUTTONBN_DEFAULT_SERVER_URL = 'http://test-install.blindsidenetworks.com/bigbluebutton/';
-/** @var BIGBLUEBUTTONBN_DEFAULT_SHARED_SECRET string of default bigbluebutton server shared secret */
-const BIGBLUEBUTTONBN_DEFAULT_SHARED_SECRET = '8cd8ef52e8e101574e400365b55e11a6';
-/** @var BIGBLUEBUTTONBN_LOG_EVENT_ADD string defines the bigbluebuttonbn Add event */
-const BIGBLUEBUTTONBN_LOG_EVENT_ADD = 'Add';
-/** @var BIGBLUEBUTTONBN_LOG_EVENT_EDIT string defines the bigbluebuttonbn Edit event */
-const BIGBLUEBUTTONBN_LOG_EVENT_EDIT = 'Edit';
-/** @var BIGBLUEBUTTONBN_LOG_EVENT_CREATE string defines the bigbluebuttonbn Create event */
-const BIGBLUEBUTTONBN_LOG_EVENT_CREATE = 'Create';
-/** @var BIGBLUEBUTTONBN_LOG_EVENT_JOIN string defines the bigbluebuttonbn Join event */
-const BIGBLUEBUTTONBN_LOG_EVENT_JOIN = 'Join';
-/** @var BIGBLUEBUTTONBN_LOG_EVENT_PLAYED string defines the bigbluebuttonbn Playback event */
-const BIGBLUEBUTTONBN_LOG_EVENT_PLAYED = 'Played';
-/** @var BIGBLUEBUTTONBN_LOG_EVENT_LOGOUT string defines the bigbluebuttonbn Logout event */
-const BIGBLUEBUTTONBN_LOG_EVENT_LOGOUT = 'Logout';
-/** @var BIGBLUEBUTTONBN_LOG_EVENT_IMPORT string defines the bigbluebuttonbn Import event */
-const BIGBLUEBUTTONBN_LOG_EVENT_IMPORT = 'Import';
-/** @var BIGBLUEBUTTONBN_LOG_EVENT_DELETE string defines the bigbluebuttonbn Delete event */
-const BIGBLUEBUTTONBN_LOG_EVENT_DELETE = 'Delete';
-/** @var BIGBLUEBUTTON_LOG_EVENT_CALLBACK string defines the bigbluebuttonbn Callback event */
-const BIGBLUEBUTTON_LOG_EVENT_CALLBACK = 'Callback';
-/** @var BIGBLUEBUTTON_LOG_EVENT_SUMMARY string defines the bigbluebuttonbn Summary event */
-const BIGBLUEBUTTON_LOG_EVENT_SUMMARY = 'Summary';
 /**
  * Indicates API features that the bigbluebuttonbn supports.
  *
@@ -97,7 +58,6 @@ function bigbluebuttonbn_supports($feature) {
         (string) FEATURE_IDNUMBER => true,
         (string) FEATURE_GROUPS => true,
         (string) FEATURE_GROUPINGS => true,
-        (string) FEATURE_GROUPMEMBERSONLY => true,
         (string) FEATURE_MOD_INTRO => true,
         (string) FEATURE_BACKUP_MOODLE2 => true,
         (string) FEATURE_COMPLETION_TRACKS_VIEWS => true,
@@ -139,7 +99,7 @@ function bigbluebuttonbn_get_completion_state($course, $cm, $userid, $type) {
 
     $sql  = "SELECT * FROM {bigbluebuttonbn_logs} ";
     $sql .= "WHERE bigbluebuttonbnid = ? AND userid = ? AND log = ?";
-    $logs = $DB->get_records_sql($sql, array($bigbluebuttonbn->id, $userid, BIGBLUEBUTTON_LOG_EVENT_SUMMARY));
+    $logs = $DB->get_records_sql($sql, array($bigbluebuttonbn->id, $userid, bbb_constants::BIGBLUEBUTTON_LOG_EVENT_SUMMARY));
 
     if ($bigbluebuttonbn->completionattendance) {
         if (!$logs) {
@@ -211,9 +171,9 @@ function bigbluebuttonbn_get_completion_state($course, $cm, $userid, $type) {
 function bigbluebuttonbn_add_instance($bigbluebuttonbn) {
     global $DB;
     // Excecute preprocess.
-    bigbluebuttonbn_process_pre_save($bigbluebuttonbn);
+    instance::bigbluebuttonbn_process_pre_save($bigbluebuttonbn);
     // Pre-set initial values.
-    $bigbluebuttonbn->presentation = bigbluebuttonbn_get_media_file($bigbluebuttonbn);
+    $bigbluebuttonbn->presentation = files::bigbluebuttonbn_get_media_file($bigbluebuttonbn);
     // Insert a record.
     $bigbluebuttonbn->id = $DB->insert_record('bigbluebuttonbn', $bigbluebuttonbn);
     // Encode meetingid.
@@ -221,9 +181,9 @@ function bigbluebuttonbn_add_instance($bigbluebuttonbn) {
     // Set the meetingid column in the bigbluebuttonbn table.
     $DB->set_field('bigbluebuttonbn', 'meetingid', $bigbluebuttonbn->meetingid, array('id' => $bigbluebuttonbn->id));
     // Log insert action.
-    bigbluebuttonbn_log($bigbluebuttonbn, BIGBLUEBUTTONBN_LOG_EVENT_ADD);
+    logs::bigbluebuttonbn_log($bigbluebuttonbn, bbb_constants::BIGBLUEBUTTONBN_LOG_EVENT_ADD);
     // Complete the process.
-    bigbluebuttonbn_process_post_save($bigbluebuttonbn);
+    instance::bigbluebuttonbn_process_post_save($bigbluebuttonbn);
     return $bigbluebuttonbn->id;
 }
 
@@ -238,18 +198,18 @@ function bigbluebuttonbn_add_instance($bigbluebuttonbn) {
 function bigbluebuttonbn_update_instance($bigbluebuttonbn) {
     global $DB;
     // Excecute preprocess.
-    bigbluebuttonbn_process_pre_save($bigbluebuttonbn);
+    instance::bigbluebuttonbn_process_pre_save($bigbluebuttonbn);
     // Pre-set initial values.
     $bigbluebuttonbn->id = $bigbluebuttonbn->instance;
-    $bigbluebuttonbn->presentation = bigbluebuttonbn_get_media_file($bigbluebuttonbn);
+    $bigbluebuttonbn->presentation = files::bigbluebuttonbn_get_media_file($bigbluebuttonbn);
     // Update a record.
     $DB->update_record('bigbluebuttonbn', $bigbluebuttonbn);
     // Get the meetingid column in the bigbluebuttonbn table.
     $bigbluebuttonbn->meetingid = (string)$DB->get_field('bigbluebuttonbn', 'meetingid', array('id' => $bigbluebuttonbn->id));
     // Log update action.
-    bigbluebuttonbn_log($bigbluebuttonbn, BIGBLUEBUTTONBN_LOG_EVENT_EDIT);
+    logs::bigbluebuttonbn_log($bigbluebuttonbn, bbb_constants::BIGBLUEBUTTONBN_LOG_EVENT_EDIT);
     // Complete the process.
-    bigbluebuttonbn_process_post_save($bigbluebuttonbn);
+    instance::bigbluebuttonbn_process_post_save($bigbluebuttonbn);
     return true;
 }
 
@@ -283,26 +243,9 @@ function bigbluebuttonbn_delete_instance($id) {
     }
 
     // Log action performed.
-    bigbluebuttonbn_delete_instance_log($bigbluebuttonbn);
+    logs::bigbluebuttonbn_delete_instance_log($bigbluebuttonbn);
 
     return $result;
-}
-
-/**
- * Given an ID of an instance of this module,
- * this function will permanently delete the data that depends on it.
- *
- * @param object $bigbluebuttonbn Id of the module instance
- *
- * @return bool Success/Failure
- */
-function bigbluebuttonbn_delete_instance_log($bigbluebuttonbn) {
-    global $DB;
-    $sql  = "SELECT * FROM {bigbluebuttonbn_logs} ";
-    $sql .= "WHERE bigbluebuttonbnid = ? AND log = ? AND ". $DB->sql_compare_text('meta') . " = ?";
-    $logs = $DB->get_records_sql($sql, array($bigbluebuttonbn->id, BIGBLUEBUTTONBN_LOG_EVENT_CREATE, "{\"record\":true}"));
-    $meta = "{\"has_recordings\":" . empty($logs) ? "true" : "false" . "}";
-    bigbluebuttonbn_log($bigbluebuttonbn, BIGBLUEBUTTONBN_LOG_EVENT_DELETE, [], $meta);
 }
 
 /**
@@ -351,7 +294,7 @@ function bigbluebuttonbn_user_complete($courseorid, $userorid, $bigbluebuttonbn)
     $sql = "SELECT COUNT(*) FROM {bigbluebuttonbn_logs} ";
     $sql .= "WHERE courseid = ? AND bigbluebuttonbnid = ? AND userid = ? AND (log = ? OR log = ?)";
     $result = $DB->count_records_sql($sql, array($course->id, $bigbluebuttonbn->id, $user->id,
-                                              BIGBLUEBUTTONBN_LOG_EVENT_JOIN, BIGBLUEBUTTONBN_LOG_EVENT_PLAYED));
+        bbb_constants::BIGBLUEBUTTONBN_LOG_EVENT_JOIN, bbb_constants::BIGBLUEBUTTONBN_LOG_EVENT_PLAYED));
     return $result;
 }
 
@@ -365,27 +308,13 @@ function bigbluebuttonbn_get_extra_capabilities() {
 }
 
 /**
- * Define items to be reset by course/reset.php
- *
- * @return array
- */
-function bigbluebuttonbn_reset_course_items() {
-    $items = array("events" => 0, "tags" => 0, "logs" => 0);
-    // Include recordings only if enabled.
-    if ((boolean)\mod_bigbluebuttonbn\locallib\config::recordings_enabled()) {
-        $items["recordings"] = 0;
-    }
-    return $items;
-}
-
-/**
  * Called by course/reset.php
  *
  * @param object $mform
  * @return void
  */
 function bigbluebuttonbn_reset_course_form_definition(&$mform) {
-    $items = bigbluebuttonbn_reset_course_items();
+    $items = reset::bigbluebuttonbn_reset_course_items();
     $mform->addElement('header', 'bigbluebuttonbnheader', get_string('modulenameplural', 'bigbluebuttonbn'));
     foreach ($items as $item => $default) {
         $mform->addElement(
@@ -407,7 +336,7 @@ function bigbluebuttonbn_reset_course_form_definition(&$mform) {
  */
 function bigbluebuttonbn_reset_course_form_defaults($course) {
     $formdefaults = array();
-    $items = bigbluebuttonbn_reset_course_items();
+    $items = reset::bigbluebuttonbn_reset_course_items();
     // All unchecked by default.
     foreach ($items as $item => $default) {
         $formdefaults["reset_bigbluebuttonbn_{$item}"] = $default;
@@ -422,179 +351,37 @@ function bigbluebuttonbn_reset_course_form_defaults($course) {
  * @return array status array
  */
 function bigbluebuttonbn_reset_userdata($data) {
-    $items = bigbluebuttonbn_reset_course_items();
+    $items = reset::bigbluebuttonbn_reset_course_items();
     $status = array();
     // Any changes to the list of dates that needs to be rolled should be same during course restore and course reset.
     // See MDL-9367.
     if (array_key_exists('recordings', $items) && !empty($data->reset_bigbluebuttonbn_recordings)) {
         // Remove all the recordings from a BBB server that are linked to the room/activities in this course.
-        bigbluebuttonbn_reset_recordings($data->courseid);
+        reset::bigbluebuttonbn_reset_recordings($data->courseid);
         unset($items['recordings']);
-        $status[] = bigbluebuttonbn_reset_getstatus('recordings');
+        $status[] = reset::bigbluebuttonbn_reset_getstatus('recordings');
     }
     if (!empty($data->reset_bigbluebuttonbn_tags)) {
         // Remove all the tags linked to the room/activities in this course.
-        bigbluebuttonbn_reset_tags($data->courseid);
+        reset::bigbluebuttonbn_reset_tags($data->courseid);
         unset($items['tags']);
-        $status[] = bigbluebuttonbn_reset_getstatus('tags');
+        $status[] = reset::bigbluebuttonbn_reset_getstatus('tags');
     }
     // TODO : seems to be duplicated code unless we just want to force reset tags.
     foreach ($items as $item => $default) {
         // Remove instances or elements linked to this course, others than recordings or tags.
         if (!empty($data->{"reset_bigbluebuttonbn_{$item}"})) {
             call_user_func("bigbluebuttonbn_reset_{$item}", $data->courseid);
-            $status[] = bigbluebuttonbn_reset_getstatus($item);
+            $status[] = reset::bigbluebuttonbn_reset_getstatus($item);
         }
     }
     return $status;
 }
 
-/**
- * Returns status used on every defined reset action.
- *
- * @param string $item
- * @return array status array
- */
-function bigbluebuttonbn_reset_getstatus($item) {
-    return array('component' => get_string('modulenameplural', 'bigbluebuttonbn')
-        , 'item' => get_string("removed{$item}", 'bigbluebuttonbn')
-        , 'error' => false);
-}
-
-/**
- * Used by the reset_course_userdata for deleting events linked to bigbluebuttonbn instances in the course.
- *
- * @param string $courseid
- * @return array status array
- */
-function bigbluebuttonbn_reset_events($courseid) {
-    global $DB;
-    // Remove all the events.
-    return $DB->delete_records('event', array('modulename' => 'bigbluebuttonbn', 'courseid' => $courseid));
-}
-
-/**
- * Used by the reset_course_userdata for deleting tags linked to bigbluebuttonbn instances in the course.
- *
- * @param array $courseid
- * @return array status array
- */
-function bigbluebuttonbn_reset_tags($courseid) {
-    global $DB;
-    // Remove all the tags linked to the room/activities in this course.
-    if ($bigbluebuttonbns = $DB->get_records('bigbluebuttonbn', array('course' => $courseid))) {
-        foreach ($bigbluebuttonbns as $bigbluebuttonbn) {
-            if (!$cm = get_coursemodule_from_instance('bigbluebuttonbn', $bigbluebuttonbn->id, $courseid)) {
-                continue;
-            }
-            $context = context_module::instance($cm->id);
-            core_tag_tag::delete_instances('mod_bigbluebuttonbn', null, $context->id);
-        }
-    }
-}
-
-/**
- * Used by the reset_course_userdata for deleting bigbluebuttonbn_logs linked to bigbluebuttonbn instances in the course.
- *
- * @param string $courseid
- * @return array status array
- */
-function bigbluebuttonbn_reset_logs($courseid) {
-    global $DB;
-    // Remove all the logs.
-    return $DB->delete_records('bigbluebuttonbn_logs', array('courseid' => $courseid));
-}
-
-/**
- * Used by the reset_course_userdata for deleting recordings in a BBB server linked to bigbluebuttonbn instances in the course.
- *
- * @param string $courseid
- * @return array status array
- */
-function bigbluebuttonbn_reset_recordings($courseid) {
-    require_once(__DIR__.'/locallib.php');
-    // Criteria for search [courseid | bigbluebuttonbn=null | subset=false | includedeleted=true].
-    $recordings = bigbluebuttonbn_get_recordings($courseid, null, false, true);
-    // Remove all the recordings.
-    bigbluebuttonbn_delete_recordings(implode(",", array_keys($recordings)));
-}
-
-/**
- * List of view style log actions.
- *
- * @return string[]
- */
-function bigbluebuttonbn_get_view_actions() {
-    return array('view', 'view all');
-}
-
-/**
- * List of update style log actions.
- *
- * @return string[]
- */
-function bigbluebuttonbn_get_post_actions() {
-    return array('update', 'add', 'delete');
-}
-
-/**
- * Print an overview of all bigbluebuttonbn instances for the courses.
- *
- * @param array $courses
- * @param array $htmlarray Passed by reference
- *
- * @return void
- */
-function bigbluebuttonbn_print_overview($courses, &$htmlarray) {
-    if (empty($courses) || !is_array($courses)) {
-        return array();
-    }
-    $bns = get_all_instances_in_courses('bigbluebuttonbn', $courses);
-    foreach ($bns as $bn) {
-        $now = time();
-        if ($bn->openingtime and (!$bn->closingtime or $bn->closingtime > $now)) {
-            // A bigbluebuttonbn is scheduled.
-            if (empty($htmlarray[$bn->course]['bigbluebuttonbn'])) {
-                $htmlarray[$bn->course]['bigbluebuttonbn'] = '';
-            }
-            // Make sure we print all bigbluebutton instances.
-            $htmlarray[$bn->course]['bigbluebuttonbn'] .= bigbluebuttonbn_print_overview_element($bn, $now);
-        }
-    }
-}
-
-/**
- * Print an overview of a bigbluebuttonbn instance.
- *
- * @param array $bigbluebuttonbn
- * @param int $now
- *
- * @return string
- */
-function bigbluebuttonbn_print_overview_element($bigbluebuttonbn, $now) {
-    global $CFG;
-    $start = 'started_at';
-    if ($bigbluebuttonbn->openingtime > $now) {
-        $start = 'starts_at';
-    }
-    $classes = '';
-    if ($bigbluebuttonbn->visible) {
-        $classes = 'class="dimmed" ';
-    }
-    $str  = '<div class="bigbluebuttonbn overview">'."\n";
-    $str .= '  <div class="name">'.get_string('modulename', 'bigbluebuttonbn').':&nbsp;'."\n";
-    $str .= '    <a '.$classes.'href="'.$CFG->wwwroot.'/mod/bigbluebuttonbn/view.php?id='.$bigbluebuttonbn->coursemodule.
-      '">'.$bigbluebuttonbn->name.'</a>'."\n";
-    $str .= '  </div>'."\n";
-    $str .= '  <div class="info">'.get_string($start, 'bigbluebuttonbn').': '.userdate($bigbluebuttonbn->openingtime).
-        '</div>'."\n";
-    if (!empty($bigbluebuttonbn->closingtime)) {
-        $str .= '  <div class="info">'.get_string('ends_at', 'bigbluebuttonbn').': '.userdate($bigbluebuttonbn->closingtime)
-                .'</div>'."\n";
-    }
-    $str .= '</div>'."\n";
-    return $str;
-}
+// Removed bigbluebuttonbn_get_view_actions as deprecated and used for legacy logs.
+// Removed bigbluebuttonbn_get_post_actions as deprecated and used for legacy logs.
+// Removed bigbluebuttonbn_print_overview as deprecated since 3.2.
+// Removed bigbluebuttonbn_print_overview_element as deprecated since 3.2.
 
 /**
  * Given a course_module object, this function returns any
@@ -613,7 +400,7 @@ function bigbluebuttonbn_get_coursemodule_info($coursemodule) {
     $fields = 'id, name, intro, introformat, completionattendance';
     $bigbluebuttonbn = $DB->get_record('bigbluebuttonbn', $dbparams, $fields);
     if (!$bigbluebuttonbn) {
-        return false;
+        return null;
     }
     $info = new cached_cm_info();
     $info->name = $bigbluebuttonbn->name;
@@ -659,253 +446,6 @@ function mod_bigbluebuttonbn_get_completion_active_rule_descriptions($cm) {
 }
 
 /**
- * Runs any processes that must run before a bigbluebuttonbn insert/update.
- *
- * @param object $bigbluebuttonbn BigBlueButtonBN form data
- *
- * @return void
- **/
-function bigbluebuttonbn_process_pre_save(&$bigbluebuttonbn) {
-    bigbluebuttonbn_process_pre_save_instance($bigbluebuttonbn);
-    bigbluebuttonbn_process_pre_save_checkboxes($bigbluebuttonbn);
-    bigbluebuttonbn_process_pre_save_common($bigbluebuttonbn);
-    $bigbluebuttonbn->participants = htmlspecialchars_decode($bigbluebuttonbn->participants);
-}
-
-/**
- * Runs process for defining the instance (insert/update).
- *
- * @param object $bigbluebuttonbn BigBlueButtonBN form data
- *
- * @return void
- **/
-function bigbluebuttonbn_process_pre_save_instance(&$bigbluebuttonbn) {
-    require_once(__DIR__.'/locallib.php');
-    $bigbluebuttonbn->timemodified = time();
-    if ((integer)$bigbluebuttonbn->instance == 0) {
-        $bigbluebuttonbn->meetingid = 0;
-        $bigbluebuttonbn->timecreated = time();
-        $bigbluebuttonbn->timemodified = 0;
-        // As it is a new activity, assign passwords.
-        $bigbluebuttonbn->moderatorpass = bigbluebuttonbn_random_password(12);
-        $bigbluebuttonbn->viewerpass = bigbluebuttonbn_random_password(12, $bigbluebuttonbn->moderatorpass);
-    }
-}
-
-/**
- * Runs process for assigning default value to checkboxes.
- *
- * @param object $bigbluebuttonbn BigBlueButtonBN form data
- *
- * @return void
- **/
-function bigbluebuttonbn_process_pre_save_checkboxes(&$bigbluebuttonbn) {
-    if (!isset($bigbluebuttonbn->wait)) {
-        $bigbluebuttonbn->wait = 0;
-    }
-    if (!isset($bigbluebuttonbn->record)) {
-        $bigbluebuttonbn->record = 0;
-    }
-    if (!isset($bigbluebuttonbn->recordallfromstart)) {
-        $bigbluebuttonbn->recordallfromstart = 0;
-    }
-    if (!isset($bigbluebuttonbn->recordhidebutton)) {
-        $bigbluebuttonbn->recordhidebutton = 0;
-    }
-    if (!isset($bigbluebuttonbn->recordings_html)) {
-        $bigbluebuttonbn->recordings_html = 0;
-    }
-    if (!isset($bigbluebuttonbn->recordings_deleted)) {
-        $bigbluebuttonbn->recordings_deleted = 0;
-    }
-    if (!isset($bigbluebuttonbn->recordings_imported)) {
-        $bigbluebuttonbn->recordings_imported = 0;
-    }
-    if (!isset($bigbluebuttonbn->recordings_preview)) {
-        $bigbluebuttonbn->recordings_preview = 0;
-    }
-    if (!isset($bigbluebuttonbn->muteonstart)) {
-        $bigbluebuttonbn->muteonstart = 0;
-    }
-    if (!isset($bigbluebuttonbn->disablecam)) {
-        $bigbluebuttonbn->disablecam = 0;
-    }
-    if (!isset($bigbluebuttonbn->disablemic)) {
-        $bigbluebuttonbn->disablemic = 0;
-    }
-    if (!isset($bigbluebuttonbn->disableprivatechat)) {
-        $bigbluebuttonbn->disableprivatechat = 0;
-    }
-    if (!isset($bigbluebuttonbn->disablepublicchat)) {
-        $bigbluebuttonbn->disablepublicchat = 0;
-    }
-    if (!isset($bigbluebuttonbn->disablenote)) {
-        $bigbluebuttonbn->disablenote = 0;
-    }
-    if (!isset($bigbluebuttonbn->hideuserlist)) {
-        $bigbluebuttonbn->hideuserlist = 0;
-    }
-    if (!isset($bigbluebuttonbn->lockedlayout)) {
-        $bigbluebuttonbn->lockedlayout = 0;
-    }
-    if (!isset($bigbluebuttonbn->lockonjoin)) {
-        $bigbluebuttonbn->lockonjoin = 0;
-    }
-    if (!isset($bigbluebuttonbn->lockonjoinconfigurable)) {
-        $bigbluebuttonbn->lockonjoinconfigurable = 0;
-    }
-    if (!isset($bigbluebuttonbn->recordings_validate_url)) {
-        $bigbluebuttonbn->recordings_validate_url = 1;
-    }
-}
-
-/**
- * Runs process for wipping common settings when 'recordings only'.
- *
- * @param object $bigbluebuttonbn BigBlueButtonBN form data
- *
- * @return void
- **/
-function bigbluebuttonbn_process_pre_save_common(&$bigbluebuttonbn) {
-    // Make sure common settings are removed when 'recordings only'.
-    if ($bigbluebuttonbn->type == BIGBLUEBUTTONBN_TYPE_RECORDING_ONLY) {
-        $bigbluebuttonbn->groupmode = 0;
-        $bigbluebuttonbn->groupingid = 0;
-    }
-}
-
-/**
- * Runs any processes that must be run after a bigbluebuttonbn insert/update.
- *
- * @param object $bigbluebuttonbn BigBlueButtonBN form data
- *
- * @return void
- **/
-function bigbluebuttonbn_process_post_save(&$bigbluebuttonbn) {
-    if (isset($bigbluebuttonbn->notification) && $bigbluebuttonbn->notification) {
-        bigbluebuttonbn_process_post_save_notification($bigbluebuttonbn);
-    }
-    bigbluebuttonbn_process_post_save_event($bigbluebuttonbn);
-    bigbluebuttonbn_process_post_save_completion($bigbluebuttonbn);
-}
-
-/**
- * Generates a message on insert/update which is sent to all users enrolled.
- *
- * @param object $bigbluebuttonbn BigBlueButtonBN form data
- *
- * @return void
- **/
-function bigbluebuttonbn_process_post_save_notification(&$bigbluebuttonbn) {
-    $action = get_string('mod_form_field_notification_msg_modified', 'bigbluebuttonbn');
-    if (isset($bigbluebuttonbn->add) && !empty($bigbluebuttonbn->add)) {
-        $action = get_string('mod_form_field_notification_msg_created', 'bigbluebuttonbn');
-    }
-    \mod_bigbluebuttonbn\locallib\notifier::notify_instance_updated($bigbluebuttonbn, $action);
-}
-
-/**
- * Generates an event after a bigbluebuttonbn insert/update.
- *
- * @param object $bigbluebuttonbn BigBlueButtonBN form data
- *
- * @return void
- **/
-function bigbluebuttonbn_process_post_save_event(&$bigbluebuttonbn) {
-    global $CFG, $DB;
-    require_once($CFG->dirroot.'/calendar/lib.php');
-    $eventid = $DB->get_field('event', 'id', array('modulename' => 'bigbluebuttonbn',
-        'instance' => $bigbluebuttonbn->id));
-    // Delete the event from calendar when/if openingtime is NOT set.
-    if (!isset($bigbluebuttonbn->openingtime) || !$bigbluebuttonbn->openingtime) {
-        if ($eventid) {
-            $calendarevent = calendar_event::load($eventid);
-            $calendarevent->delete();
-        }
-        return;
-    }
-    // Add evento to the calendar as openingtime is set.
-    $event = new stdClass();
-    $event->eventtype = BIGBLUEBUTTON_EVENT_MEETING_START;
-    $event->type = CALENDAR_EVENT_TYPE_ACTION;
-    $event->name = get_string('calendarstarts', 'bigbluebuttonbn', $bigbluebuttonbn->name);
-    $event->description = format_module_intro('bigbluebuttonbn', $bigbluebuttonbn, $bigbluebuttonbn->coursemodule, false);
-    $event->format      = FORMAT_HTML;
-    $event->courseid = $bigbluebuttonbn->course;
-    $event->groupid = 0;
-    $event->userid = 0;
-    $event->modulename = 'bigbluebuttonbn';
-    $event->instance = $bigbluebuttonbn->id;
-    $event->timestart = $bigbluebuttonbn->openingtime;
-    $event->timeduration = 0;
-    $event->timesort = $event->timestart;
-    $event->visible = instance_is_visible('bigbluebuttonbn', $bigbluebuttonbn);
-    $event->priority = null;
-    // Update the event in calendar when/if eventid was found.
-    if ($eventid) {
-        $event->id = $eventid;
-        $calendarevent = calendar_event::load($eventid);
-        $calendarevent->update($event);
-        return;
-    }
-    calendar_event::create($event);
-}
-
-/**
- * Generates an event after a bigbluebuttonbn activity is completed.
- *
- * @param object $bigbluebuttonbn BigBlueButtonBN form data
- *
- * @return void
- **/
-function bigbluebuttonbn_process_post_save_completion($bigbluebuttonbn) {
-    if (!empty($bigbluebuttonbn->completionexpected)) {
-        \core_completion\api::update_completion_date_event(
-            $bigbluebuttonbn->coursemodule,
-            'bigbluebuttonbn',
-            $bigbluebuttonbn->id,
-            $bigbluebuttonbn->completionexpected
-          );
-    }
-}
-
-/**
- * Get a full path to the file attached as a preuploaded presentation
- * or if there is none, set the presentation field will be set to blank.
- *
- * @param object $bigbluebuttonbn BigBlueButtonBN form data
- *
- * @return string
- */
-function bigbluebuttonbn_get_media_file(&$bigbluebuttonbn) {
-    if (!isset($bigbluebuttonbn->presentation) || $bigbluebuttonbn->presentation == '') {
-        return '';
-    }
-    $context = context_module::instance($bigbluebuttonbn->coursemodule);
-    // Set the filestorage object.
-    $fs = get_file_storage();
-    // Save the file if it exists that is currently in the draft area.
-    file_save_draft_area_files($bigbluebuttonbn->presentation, $context->id, 'mod_bigbluebuttonbn', 'presentation', 0);
-    // Get the file if it exists.
-    $files = $fs->get_area_files(
-        $context->id,
-        'mod_bigbluebuttonbn',
-        'presentation',
-        0,
-        'itemid, filepath, filename',
-        false
-    );
-    // Check that there is a file to process.
-    $filesrc = '';
-    if (count($files) == 1) {
-        // Get the first (and only) file.
-        $file = reset($files);
-        $filesrc = '/'.$file->get_filename();
-    }
-    return $filesrc;
-}
-
-/**
  * Serves the bigbluebuttonbn attachments. Implements needed access control ;-).
  *
  * @category files
@@ -921,171 +461,15 @@ function bigbluebuttonbn_get_media_file(&$bigbluebuttonbn) {
  * @return false|null false if file not found, does not return if found - justsend the file
  */
 function bigbluebuttonbn_pluginfile($course, $cm, $context, $filearea, $args, $forcedownload, array $options = array()) {
-    if (!bigbluebuttonbn_pluginfile_valid($context, $filearea)) {
+    if (!files::bigbluebuttonbn_pluginfile_valid($context, $filearea)) {
         return false;
     }
-    $file = bigbluebuttonbn_pluginfile_file($course, $cm, $context, $filearea, $args);
+    $file = files::bigbluebuttonbn_pluginfile_file($course, $cm, $context, $filearea, $args);
     if (empty($file)) {
         return false;
     }
     // Finally send the file.
-    send_stored_file($file, 0, 0, $forcedownload, $options); // download MUST be forced - security!
-}
-
-/**
- * Helper for validating pluginfile.
- * @param stdClass $context       context object
- * @param string   $filearea      file area
- *
- * @return false|null false if file not valid
- */
-function bigbluebuttonbn_pluginfile_valid($context, $filearea) {
-
-    // Can be in context module or in context_system (if is the presentation by default).
-    if (!in_array($context->contextlevel, array(CONTEXT_MODULE, CONTEXT_SYSTEM))) {
-        return false;
-    }
-
-    if (!array_key_exists($filearea, bigbluebuttonbn_get_file_areas())) {
-        return false;
-    }
-
-    return true;
-}
-
-/**
- * Helper for getting pluginfile.
- *
- * @param stdClass $course        course object
- * @param stdClass $cm            course module object
- * @param stdClass $context       context object
- * @param string   $filearea      file area
- * @param array    $args          extra arguments
- *
- * @return object
- */
-function bigbluebuttonbn_pluginfile_file($course, $cm, $context, $filearea, $args) {
-    $filename = bigbluebuttonbn_pluginfile_filename($course, $cm, $context, $args);
-    if (!$filename) {
-        return false;
-    }
-    $fullpath = "/$context->id/mod_bigbluebuttonbn/$filearea/0/".$filename;
-    $fs = get_file_storage();
-    $file = $fs->get_file_by_hash(sha1($fullpath));
-    if (!$file || $file->is_directory()) {
-        return false;
-    }
-    return $file;
-}
-
-/**
- * Helper for give access to the file configured in setting as default presentation.
- *
- * @param stdClass $course        course object
- * @param stdClass $cm            course module object
- * @param stdClass $context       context object
- * @param array    $args          extra arguments
- *
- * @return array
- */
-function bigbluebuttonbn_default_presentation_get_file($course, $cm, $context, $args) {
-
-    // The difference with the standard bigbluebuttonbn_pluginfile_filename() are.
-    // - Context is system, so we don't need to check the cmid in this case.
-    // - The area is "presentationdefault_cache".
-    if (count($args) > 1) {
-        $cache = cache::make_from_params(
-            cache_store::MODE_APPLICATION,
-            'mod_bigbluebuttonbn',
-            'presentationdefault_cache'
-        );
-
-        $noncekey = sha1($context->id);
-        $presentationnonce = $cache->get($noncekey);
-        $noncevalue = $presentationnonce['value'];
-        $noncecounter = $presentationnonce['counter'];
-        if ($args['0'] != $noncevalue) {
-            return;
-        }
-
-        // The nonce value is actually used twice because BigBlueButton reads the file two times.
-        $noncecounter += 1;
-        $cache->set($noncekey, array('value' => $noncevalue, 'counter' => $noncecounter));
-        if ($noncecounter == 2) {
-            $cache->delete($noncekey);
-        }
-        return($args['1']);
-    }
-    require_course_login($course, true, $cm);
-    if (!has_capability('mod/bigbluebuttonbn:join', $context)) {
-        return;
-    }
-    return implode('/', $args);
-}
-
-/**
- * Helper for getting pluginfile name.
- *
- * @param stdClass $course        course object
- * @param stdClass $cm            course module object
- * @param stdClass $context       context object
- * @param array    $args          extra arguments
- *
- * @return array
- */
-function bigbluebuttonbn_pluginfile_filename($course, $cm, $context, $args) {
-    global $DB;
-
-    if ($context->contextlevel == CONTEXT_SYSTEM) {
-        // Plugin has a file to use as default in general setting.
-        return(bigbluebuttonbn_default_presentation_get_file($course, $cm, $context, $args));
-    }
-
-    if (count($args) > 1) {
-        if (!$bigbluebuttonbn = $DB->get_record('bigbluebuttonbn', array('id' => $cm->instance))) {
-            return;
-        }
-        $cache = cache::make_from_params(cache_store::MODE_APPLICATION, 'mod_bigbluebuttonbn', 'presentation_cache');
-        $noncekey = sha1($bigbluebuttonbn->id);
-        $presentationnonce = $cache->get($noncekey);
-        if (!empty($presentationnonce)) {
-            $noncevalue = $presentationnonce['value'];
-            $noncecounter = $presentationnonce['counter'];
-        } else {
-            $noncevalue = null;
-            $noncecounter = 0;
-        }
-
-        if ($args['0'] != $noncevalue) {
-            return;
-        }
-        // The nonce value is actually used twice because BigBlueButton reads the file two times.
-        $noncecounter += 1;
-        $cache->set($noncekey, array('value' => $noncevalue, 'counter' => $noncecounter));
-        if ($noncecounter == 2) {
-            $cache->delete($noncekey);
-        }
-        return $args['1'];
-    }
-    require_course_login($course, true, $cm);
-    if (!has_capability('mod/bigbluebuttonbn:join', $context)) {
-        return;
-    }
-    return implode('/', $args);
-}
-
-/**
- * Returns an array of file areas.
- *
- * @category files
- *
- * @return array a list of available file areas
- */
-function bigbluebuttonbn_get_file_areas() {
-    $areas = array();
-    $areas['presentation'] = get_string('mod_form_block_presentation', 'bigbluebuttonbn');
-    $areas['presentationdefault'] = get_string('mod_form_block_presentation_default', 'bigbluebuttonbn');
-    return $areas;
+    return send_stored_file($file, 0, 0, $forcedownload, $options); // download MUST be forced - security!
 }
 
 /**
@@ -1183,46 +567,15 @@ function mod_bigbluebuttonbn_core_calendar_provide_event_action(
 
     // Action data.
     $string = get_string('view_room', 'bigbluebuttonbn');
-    $url = new \moodle_url('/mod/bigbluebuttonbn/view.php', array('id' => $cm->id));
+    $url = new moodle_url('/mod/bigbluebuttonbn/view.php', array('id' => $cm->id));
     if (groups_get_activity_groupmode($cm) == NOGROUPS) {
         // No groups mode.
         $string = get_string('view_conference_action_join', 'bigbluebuttonbn');
-        $url = new \moodle_url('/mod/bigbluebuttonbn/bbb_view.php', array('action' => 'join',
+        $url = new moodle_url('/mod/bigbluebuttonbn/bbb_view.php', array('action' => 'join',
             'id' => $cm->id, 'bn' => $bigbluebuttonbn->id, 'timeline' => 1));
     }
 
     return $factory->create_instance($string, $url, 1, $actionable);
-}
-
-/**
- * Register a bigbluebuttonbn event
- *
- * @param object $bigbluebuttonbn
- * @param string $event
- * @param array  $overrides
- * @param string $meta
- *
- * @return bool Success/Failure
- */
-function bigbluebuttonbn_log($bigbluebuttonbn, $event, array $overrides = [], $meta = null) {
-    global $DB, $USER;
-    $log = new stdClass();
-    // Default values.
-    $log->courseid = $bigbluebuttonbn->course;
-    $log->bigbluebuttonbnid = $bigbluebuttonbn->id;
-    $log->userid = $USER->id;
-    $log->meetingid = $bigbluebuttonbn->meetingid;
-    $log->timecreated = time();
-    $log->log = $event;
-    $log->meta = $meta;
-    // Overrides.
-    foreach ($overrides as $key => $value) {
-        $log->$key = $value;
-    }
-    if (!$DB->insert_record('bigbluebuttonbn_logs', $log)) {
-        return false;
-    }
-    return true;
 }
 
 /**
