+++ tests/lib_test.php
@@ -23,9 +23,13 @@
  * @author    Laurent David (laurent@call-learning.fr)
  */
 
+use mod_bigbluebuttonbn\local\bbb_constants;
+use mod_bigbluebuttonbn\local\helpers\logs;
+
 defined('MOODLE_INTERNAL') || die();
 global $CFG;
 
+require_once($CFG->dirroot . '/mod/bigbluebuttonbn/tests/helpers.php');
 require_once($CFG->dirroot . '/mod/bigbluebuttonbn/lib.php');
 
 /**
@@ -36,71 +40,7 @@ require_once($CFG->dirroot . '/mod/bigbluebuttonbn/lib.php');
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  * @author    Laurent David (laurent@call-learning.fr)
  */
-class mod_bigbluebuttonbn_lib_testcase extends advanced_testcase {
-    /**
-     * @var testing_data_generator|null $generator
-     */
-    public $generator = null;
-    /**
-     * @var object|null $bbactivity
-     */
-    public $bbactivity = null;
-    /**
-     * @var object|null $course
-     */
-    public $course = null;
-
-    /**
-     * Convenience function to create a instance of an bigbluebuttonactivty.
-     *
-     * @param object|null $course course to add the module to
-     * @param array $params Array of parameters to pass to the generator
-     * @param array $options Array of options to pass to the generator
-     * @return array($context, $cm, $instance) Testable wrapper around the assign class.
-     * @throws coding_exception
-     * @throws moodle_exception
-     */
-    protected function create_instance($course = null, $params = [], $options = []) {
-        if (!$course) {
-            $course = $this->course;
-        }
-        $params['course'] = $course->id;
-        $options['visible'] = 1;
-        $instance = $this->generator->create_module('bigbluebuttonbn', $params, $options);
-        list($course, $cm) = get_course_and_cm_from_instance($instance, 'bigbluebuttonbn');
-        $context = context_module::instance($cm->id);
-
-        return array($context, $cm, $instance);
-    }
-
-    /**
-     * Get the corresponding form data
-     *
-     * @param object $bbactivity the current bigbluebutton activity
-     * @param object|null $course the course or null (taken from $this->course if null)
-     * @return mixed
-     * @throws coding_exception
-     */
-    protected function get_form_data_from_instance($bbactivity, $course = null) {
-        global $USER;
-        if (!$course) {
-            $course = $this->course;
-        }
-        $currentuser = $USER;
-        $this->setAdminUser();
-        $bbactivitycm = get_coursemodule_from_instance('bigbluebuttonbn', $bbactivity->id);
-        list($cm, $context, $module, $data, $cw) = get_moduleinfo_data($bbactivitycm, $course);
-        $this->setUser($USER);
-        return $data;
-    }
-
-    public function setUp(): void {
-        global $CFG;
-        parent::setUp();
-        set_config('enablecompletion', true); // Enable completion for all tests.
-        $this->generator = $this->getDataGenerator();
-        $this->course = $this->generator->create_course(['enablecompletion' => 1]);
-    }
+class mod_bigbluebuttonbn_lib_testcase extends \bbb_simple_testcase {
 
     public function test_bigbluebuttonbn_supports() {
         $this->resetAfterTest();
@@ -124,8 +64,8 @@ class mod_bigbluebuttonbn_lib_testcase extends advanced_testcase {
         // Add a couple of fake logs.
         $overrides = array('meetingid' => $bbactivity->meetingid);
         $meta = '{"origin":0, "data": {"duration": 120, "engagement": {"chats": 2, "talks":2} }}';
-        bigbluebuttonbn_log($bbactivity, BIGBLUEBUTTON_LOG_EVENT_SUMMARY, $overrides, $meta);
-        bigbluebuttonbn_log($bbactivity, BIGBLUEBUTTON_LOG_EVENT_SUMMARY, $overrides, $meta);
+        logs::bigbluebuttonbn_log($bbactivity, bbb_constants::BIGBLUEBUTTON_LOG_EVENT_SUMMARY, $overrides, $meta);
+        logs::bigbluebuttonbn_log($bbactivity, bbb_constants::BIGBLUEBUTTON_LOG_EVENT_SUMMARY, $overrides, $meta);
         $result = bigbluebuttonbn_get_completion_state($this->course, $bbactivitycm, $user->id, COMPLETION_AND);
         $this->assertEquals(COMPLETION_AND, $result);
     }
@@ -153,15 +93,6 @@ class mod_bigbluebuttonbn_lib_testcase extends advanced_testcase {
         $this->assertTrue($result);
     }
 
-    public function test_bigbluebuttonbn_delete_instance_log() {
-        global $DB;
-        $this->resetAfterTest();
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance();
-        bigbluebuttonbn_delete_instance_log($bbactivity);
-        $this->assertTrue($DB->record_exists('bigbluebuttonbn_logs', array('bigbluebuttonbnid' => $bbactivity->id,
-                'log' => BIGBLUEBUTTONBN_LOG_EVENT_DELETE)));
-    }
-
     public function test_bigbluebuttonbn_user_outline() {
         $this->resetAfterTest();
         $user = $this->generator->create_user();
@@ -173,8 +104,8 @@ class mod_bigbluebuttonbn_lib_testcase extends advanced_testcase {
         // Now create a couple of logs.
         $overrides = array('meetingid' => $bbactivity->meetingid);
         $meta = '{"origin":0}';
-        bigbluebuttonbn_log($bbactivity, BIGBLUEBUTTONBN_LOG_EVENT_JOIN, $overrides, $meta);
-        bigbluebuttonbn_log($bbactivity, BIGBLUEBUTTONBN_LOG_EVENT_PLAYED, $overrides);
+        logs::bigbluebuttonbn_log($bbactivity, bbb_constants::BIGBLUEBUTTONBN_LOG_EVENT_JOIN, $overrides, $meta);
+        logs::bigbluebuttonbn_log($bbactivity, bbb_constants::BIGBLUEBUTTONBN_LOG_EVENT_PLAYED, $overrides);
         $result = bigbluebuttonbn_user_outline($this->course, $user, null, $bbactivity);
         $this->assertRegExp('/.* has joined the session for 2 times/', $result);
     }
@@ -186,8 +117,8 @@ class mod_bigbluebuttonbn_lib_testcase extends advanced_testcase {
         $this->setUser($user);
         $overrides = array('meetingid' => $bbactivity->meetingid);
         $meta = '{"origin":0}';
-        bigbluebuttonbn_log($bbactivity, BIGBLUEBUTTONBN_LOG_EVENT_JOIN, $overrides, $meta);
-        bigbluebuttonbn_log($bbactivity, BIGBLUEBUTTONBN_LOG_EVENT_PLAYED, $overrides);
+        logs::bigbluebuttonbn_log($bbactivity, bbb_constants::BIGBLUEBUTTONBN_LOG_EVENT_JOIN, $overrides, $meta);
+        logs::bigbluebuttonbn_log($bbactivity, bbb_constants::BIGBLUEBUTTONBN_LOG_EVENT_PLAYED, $overrides);
         $result = bigbluebuttonbn_user_complete($this->course, $user, $bbactivity);
         $this->assertEquals(2, $result);
     }
@@ -195,17 +126,6 @@ class mod_bigbluebuttonbn_lib_testcase extends advanced_testcase {
     public function test_bigbluebuttonbn_get_extra_capabilities() {
         $this->resetAfterTest();
         $this->assertEquals(array('moodle/site:accessallgroups'), bigbluebuttonbn_get_extra_capabilities());
-    }
-
-    public function test_bigbluebuttonbn_reset_course_items() {
-        global $CFG;
-        $this->resetAfterTest();
-        $CFG->bigbluebuttonbn_recordings_enabled = false;
-        $results = bigbluebuttonbn_reset_course_items();
-        $this->assertEquals(array("events" => 0, "tags" => 0, "logs" => 0), $results);
-        $CFG->bigbluebuttonbn_recordings_enabled = true;
-        $results = bigbluebuttonbn_reset_course_items();
-        $this->assertEquals(array("events" => 0, "tags" => 0, "logs" => 0, "recordings" => 0), $results);
     }
 
     public function test_bigbluebuttonbn_reset_course_form_definition() {
@@ -260,116 +180,6 @@ class mod_bigbluebuttonbn_lib_testcase extends advanced_testcase {
         ), $results[0]);
     }
 
-    public function test_bigbluebuttonbn_reset_getstatus() {
-        $this->resetAfterTest();
-        $result = bigbluebuttonbn_reset_getstatus('events');
-        $this->assertEquals(array(
-                'component' => 'BigBlueButtonBN',
-                'item' => 'Deleted events',
-                'error' => false,
-        ), $result);
-    }
-
-    public function test_bigbluebuttonbn_reset_events() {
-        global $DB;
-        $this->resetAfterTest();
-        $this->setAdminUser();
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance(
-                null,
-                ['openingtime' => time()]
-        );
-        $formdata = $this->get_form_data_from_instance($bbactivity);
-        bigbluebuttonbn_process_post_save_event($formdata);
-        $this->assertEquals(1, $DB->count_records(
-                'event',
-                array('modulename' => 'bigbluebuttonbn', 'courseid' => $this->course->id)));
-        bigbluebuttonbn_reset_events($this->course->id);
-        $this->assertEquals(0, $DB->count_records(
-                'event',
-                array('modulename' => 'bigbluebuttonbn', 'courseid' => $this->course->id)));
-    }
-
-    public function test_bigbluebuttonbn_reset_tags() {
-        $this->resetAfterTest();
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance(null,
-                array('course' => $this->course->id),
-                ['visible' => true]
-        );
-        core_tag_tag::add_item_tag('mod_bigbluebuttonbn', 'bbitem', $bbactivity->id, $bbactivitycontext, 'newtag');
-        $alltags = core_tag_tag::get_item_tags('mod_bigbluebuttonbn', 'bbitem', $bbactivity->id);
-        $this->assertCount(1, $alltags);
-        bigbluebuttonbn_reset_tags($this->course->id);
-        $alltags = core_tag_tag::get_item_tags('mod_bigbluebuttonbn', 'bbitem', $bbactivity->id);
-        $this->assertCount(0, $alltags);
-    }
-
-    public function test_bigbluebuttonbn_reset_logs() {
-        global $DB;
-        $this->resetAfterTest();
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance(null,
-                array('course' => $this->course->id),
-                ['visible' => true]
-        );
-
-        // User has already joined the meeting (there is log event BIGBLUEBUTTONBN_LOG_EVENT_JOIN already for this user).
-        $overrides = array('meetingid' => $bbactivity->meetingid);
-        $meta = '{"origin":0}';
-        bigbluebuttonbn_log($bbactivity, BIGBLUEBUTTONBN_LOG_EVENT_JOIN, $overrides, $meta);
-
-        bigbluebuttonbn_reset_logs($this->course->id);
-        $this->assertEquals(0, $DB->count_records(
-                'bigbluebuttonbn_logs',
-                array('bigbluebuttonbnid' => $bbactivity->id, 'courseid' => $this->course->id)));
-    }
-
-    public function test_bigbluebuttonbn_reset_recordings() {
-        $this->resetAfterTest();
-        // TODO complete this test.
-        $this->markTestSkipped(
-                'For now this test relies on an API call so we need to mock the API CALL.'
-        );
-    }
-
-    public function test_bigbluebuttonbn_get_view_actions() {
-        $this->resetAfterTest();
-        $this->assertEquals(array('view', 'view all'), bigbluebuttonbn_get_view_actions());
-    }
-
-    public function test_bigbluebuttonbn_get_post_actions() {
-        $this->resetAfterTest();
-        $this->assertEquals(array('update', 'add', 'delete'), bigbluebuttonbn_get_post_actions());
-    }
-
-    public function test_bigbluebuttonbn_print_overview() {
-        $this->resetAfterTest();
-
-        $this->setAdminUser(); // If not modules won't be visible.
-        list($bbactivitycontext, $bbactivitycm, $bbactivity1) = $this->create_instance(null,
-                array('course' => $this->course->id, 'openingtime' => time()),
-                ['visible' => true]
-        );
-
-        list($bbactivitycontext, $bbactivitycm, $bbactivity2) = $this->create_instance(null,
-                array('course' => $this->course->id, 'openingtime' => time()),
-                ['visible' => true]
-        );
-
-        $htmlarray = [];
-        bigbluebuttonbn_print_overview([$this->course->id => $this->course], $htmlarray);
-        $this->assertRegExp("/BigBlueButtonBN (1|2)/", $htmlarray[$this->course->id]['bigbluebuttonbn']);
-    }
-
-    public function test_bigbluebuttonbn_print_overview_element() {
-        $this->resetAfterTest();
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance();
-        // We tweak the record as it should also contain all fields from the activity instance.
-        /* @var cm_info $bbactivitycm */
-        $cmrecord = (object) array_merge((array) $bbactivity, (array) $bbactivitycm->get_course_module_record());
-        $cmrecord->coursemodule = $bbactivity->id;
-        $str = bigbluebuttonbn_print_overview_element($cmrecord, time());
-        $this->assertRegExp("/bigbluebuttonbn overview/", $str);
-    }
-
     public function test_bigbluebuttonbn_get_coursemodule_info() {
         $this->resetAfterTest();
         list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance();
@@ -410,150 +220,6 @@ class mod_bigbluebuttonbn_lib_testcase extends advanced_testcase {
 
     }
 
-    public function test_bigbluebuttonbn_process_pre_save() {
-        $this->resetAfterTest();
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance();
-        $bbformdata = $this->get_form_data_from_instance($bbactivity);
-        $bbformdata->participants = '<p>this -&gt; &quot;</p>\n';
-        $bbformdata->timemodified = time();
-        bigbluebuttonbn_process_pre_save($bbformdata);
-        $this->assertTrue($bbformdata->timemodified != 0);
-        $this->assertEquals('<p>this -> "</p>\n', $bbformdata->participants);
-    }
-
-    public function test_bigbluebuttonbn_process_pre_save_instance() {
-        $this->resetAfterTest();
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance();
-        $bbformdata = $this->get_form_data_from_instance($bbactivity);
-        $bbformdata->instance = 0;
-        $bbformdata->timemodified = time();
-        bigbluebuttonbn_process_pre_save_instance($bbformdata);
-        $this->assertTrue($bbformdata->timemodified == 0);
-    }
-
-    public function test_bigbluebuttonbn_process_pre_save_checkboxes() {
-        $this->resetAfterTest();
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance();
-        $bbformdata = $this->get_form_data_from_instance($bbactivity);
-        unset($bbformdata->wait);
-        unset($bbformdata->recordallfromstart);
-        bigbluebuttonbn_process_pre_save_checkboxes($bbformdata);
-        $this->assertTrue(isset($bbformdata->wait));
-        $this->assertTrue(isset($bbformdata->recordallfromstart));
-    }
-
-    public function test_bigbluebuttonbn_process_pre_save_common() {
-        global $CFG;
-        require_once($CFG->dirroot . '/mod/bigbluebuttonbn/locallib.php');
-        $this->resetAfterTest();
-
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) =
-                $this->create_instance(null, ['type' => BIGBLUEBUTTONBN_TYPE_RECORDING_ONLY]);
-        $bbformdata = $this->get_form_data_from_instance($bbactivity);
-
-        $bbformdata->groupmode = '1';
-        bigbluebuttonbn_process_pre_save_common($bbformdata);
-        $this->assertEquals(0, $bbformdata->groupmode);
-    }
-
-    public function test_bigbluebuttonbn_process_post_save() {
-        global $CFG;
-        require_once($CFG->dirroot . '/mod/bigbluebuttonbn/locallib.php');
-        $this->resetAfterTest();
-
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) =
-                $this->create_instance(null, ['type' => BIGBLUEBUTTONBN_TYPE_RECORDING_ONLY]);
-        $bbformdata = $this->get_form_data_from_instance($bbactivity);
-
-        // Enrol users in a course so he will receive the message.
-        $teacher = $this->generator->create_user(['role' => 'editingteacher']);
-        $this->generator->enrol_user($teacher->id, $this->course->id);
-
-        // Mark the form to trigger notification.
-        $bbformdata->notification = true;
-        $messagesink = $this->redirectMessages();
-        bigbluebuttonbn_process_post_save($bbformdata);
-        // Now run cron.
-        ob_start();
-        $this->runAdhocTasks();
-        ob_get_clean(); // Suppress output as it can fail the test.
-        $this->assertEquals(1, $messagesink->count());
-    }
-
-    public function test_bigbluebuttonbn_process_post_save_notification() {
-        global $CFG;
-        require_once($CFG->dirroot . '/mod/bigbluebuttonbn/locallib.php');
-        $this->resetAfterTest();
-
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) =
-                $this->create_instance(null, ['type' => BIGBLUEBUTTONBN_TYPE_RECORDING_ONLY]);
-        $bbformdata = $this->get_form_data_from_instance($bbactivity);
-        $bbformdata->add = "1";
-        $messagesink = $this->redirectMessages();
-        // Enrol users in a course so he will receive the message.
-        $teacher = $this->generator->create_user(['role' => 'editingteacher']);
-        $this->generator->enrol_user($teacher->id, $this->course->id);
-
-        bigbluebuttonbn_process_post_save_notification($bbformdata);
-        // Now run cron.
-        ob_start();
-        $this->runAdhocTasks();
-        ob_get_clean(); // Suppress output as it can fail the test.
-        $this->assertEquals(1, $messagesink->count());
-    }
-
-    public function test_bigbluebuttonbn_process_post_save_event() {
-        $this->resetAfterTest();
-        $this->setAdminUser();
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance();
-        $eventsink = $this->redirectEvents();
-        $bbformdata = $this->get_form_data_from_instance($bbactivity);
-        $bbformdata->openingtime = time();
-        bigbluebuttonbn_process_post_save_event($bbformdata);
-        $this->assertNotEmpty($eventsink->get_events());
-    }
-
-    public function test_bigbluebuttonbn_process_post_save_completion() {
-        $this->resetAfterTest();
-        $this->setAdminUser();
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance();
-        $bbformdata = $this->get_form_data_from_instance($bbactivity);
-        $eventsink = $this->redirectEvents();
-        $bbformdata->completionexpected = 1;
-        bigbluebuttonbn_process_post_save_completion($bbformdata);
-        $this->assertNotEmpty($eventsink->get_events());
-    }
-
-    public function test_bigbluebuttonbn_get_media_file() {
-        $this->resetAfterTest();
-        $user = $this->generator->create_user();
-        $this->setUser($user);
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance();
-        $bbformdata = $this->get_form_data_from_instance($bbactivity);
-        $mediafilepath = bigbluebuttonbn_get_media_file($bbformdata);
-        $this->assertEmpty($mediafilepath);
-
-        // From test_delete_original_file_from_draft (lib/test/filelib_test.php)
-        // Create a bbb private file.
-        $bbbfilerecord = new stdClass;
-        $bbbfilerecord->contextid = context_module::instance($bbformdata->coursemodule)->id;
-        $bbbfilerecord->component = 'mod_bigbluebuttonbn';
-        $bbbfilerecord->filearea = 'presentation';
-        $bbbfilerecord->itemid = 0;
-        $bbbfilerecord->filepath = '/';
-        $bbbfilerecord->filename = 'bbfile.pptx';
-        $bbbfilerecord->source = 'test';
-        $fs = get_file_storage();
-        $bbbfile = $fs->create_file_from_string($bbbfilerecord, 'Presentation file content');
-        file_prepare_draft_area($bbformdata->presentation,
-                context_module::instance($bbformdata->coursemodule)->id,
-                'mod_bigbluebuttonbn',
-                'presentation', 0);
-
-        $mediafilepath = bigbluebuttonbn_get_media_file($bbformdata);
-        $this->assertEquals('/bbfile.pptx', $mediafilepath);
-    }
-
     public function test_bigbluebuttonbn_pluginfile() {
         $this->resetAfterTest();
         $this->markTestSkipped(
@@ -565,82 +231,6 @@ class mod_bigbluebuttonbn_lib_testcase extends advanced_testcase {
                'presentation', ['bbfile.pptx'], false, ['preview'=>true, 'dontdie'=>true]);
                $this->assertEquals('/bbfile.pptx', $mediafilepath);
         */
-    }
-
-    public function test_bigbluebuttonbn_pluginfile_valid() {
-        $this->resetAfterTest();
-        $this->assertFalse(bigbluebuttonbn_pluginfile_valid(context_course::instance($this->course->id), 'presentation'));
-        $this->assertTrue(bigbluebuttonbn_pluginfile_valid(context_system::instance(), 'presentation'));
-        $this->assertFalse(bigbluebuttonbn_pluginfile_valid(context_system::instance(), 'otherfilearea'));
-    }
-
-    public function test_bigbluebuttonbn_pluginfile_file() {
-        $this->resetAfterTest();
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance();
-        $user = $this->generator->create_user();
-        $this->setUser($user);
-        $this->generator->enrol_user($user->id, $this->course->id, 'editingteacher');
-        // From test_delete_original_file_from_draft (lib/test/filelib_test.php)
-        // Create a bbb private file.
-        $bbformdata = $this->get_form_data_from_instance($bbactivity);
-        $context = context_module::instance($bbformdata->coursemodule);
-
-        $bbbfilerecord = new stdClass;
-        $bbbfilerecord->contextid = $context->id;
-        $bbbfilerecord->component = 'mod_bigbluebuttonbn';
-        $bbbfilerecord->filearea = 'presentation';
-        $bbbfilerecord->itemid = 0;
-        $bbbfilerecord->filepath = '/';
-        $bbbfilerecord->filename = 'bbfile.pptx';
-        $bbbfilerecord->source = 'test';
-        $fs = get_file_storage();
-        $bbbfile = $fs->create_file_from_string($bbbfilerecord, 'Presentation file content');
-        file_prepare_draft_area($bbformdata->presentation,
-                context_module::instance($bbformdata->coursemodule)->id,
-                'mod_bigbluebuttonbn',
-                'presentation', 0);
-        list($course, $bbactivitycmuser) = get_course_and_cm_from_instance($bbactivity->id, 'bigbluebuttonbn');
-        /** @var stored_file $mediafile */
-        $mediafile = bigbluebuttonbn_pluginfile_file($this->course, $bbactivitycmuser, $context, 'presentation', ['bbfile.pptx']);
-        $this->assertEquals('bbfile.pptx', $mediafile->get_filename());
-    }
-
-    public function test_bigbluebuttonbn_default_presentation_get_file() {
-        $this->resetAfterTest();
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance();
-        $user = $this->generator->create_user();
-        $this->setUser($user);
-        $this->generator->enrol_user($user->id, $this->course->id, 'editingteacher');
-        $bbformdata = $this->get_form_data_from_instance($bbactivity);
-        // From test_delete_original_file_from_draft (lib/test/filelib_test.php)
-        // Create a bbb private file.
-        $context = context_module::instance($bbformdata->coursemodule);
-        list($course, $bbactivitycmuser) = get_course_and_cm_from_instance($bbactivity->id, 'bigbluebuttonbn');
-        $mediafile = bigbluebuttonbn_default_presentation_get_file($this->course, $bbactivitycmuser, $context, ['presentation'],
-                '/bbfile.pptx');
-        $this->assertEquals('presentation', $mediafile);
-    }
-
-    public function test_bigbluebuttonbn_pluginfile_filename() {
-        $this->resetAfterTest();
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance();
-        $user = $this->generator->create_user();
-        $this->setUser($user);
-        $this->generator->enrol_user($user->id, $this->course->id, 'editingteacher');
-        $cache = cache::make_from_params(cache_store::MODE_APPLICATION, 'mod_bigbluebuttonbn', 'presentation_cache');
-        $noncekey = sha1($bbactivity->id);
-        $presentationnonce = $cache->get($noncekey);
-        $filename = bigbluebuttonbn_pluginfile_filename($this->course, $bbactivitycm, $bbactivitycontext,
-                [$presentationnonce, 'bbfile.pptx']);
-        $this->assertEquals('bbfile.pptx', $filename);
-    }
-
-    public function test_bigbluebuttonbn_get_file_areas() {
-        $this->resetAfterTest();
-        $this->assertEquals(array(
-                'presentation' => 'Presentation content',
-                'presentationdefault' => 'Presentation default content',
-        ), bigbluebuttonbn_get_file_areas());
     }
 
     public function test_bigbluebuttonbn_view() {
@@ -703,7 +293,7 @@ class mod_bigbluebuttonbn_lib_testcase extends advanced_testcase {
         // User has already joined the meeting (there is log event BIGBLUEBUTTONBN_LOG_EVENT_JOIN already for this user).
         $overrides = array('meetingid' => $bbactivity->meetingid);
         $meta = '{"origin":0}';
-        bigbluebuttonbn_log($bbactivity, BIGBLUEBUTTONBN_LOG_EVENT_JOIN, $overrides, $meta);
+        logs::bigbluebuttonbn_log($bbactivity, bbb_constants::BIGBLUEBUTTONBN_LOG_EVENT_JOIN, $overrides, $meta);
         $bbactivity->closingtime = time() - 1000;
         $bbactivity->openingtime = time() - 2000;
         $DB->update_record('bigbluebuttonbn', $bbactivity);
@@ -732,14 +322,6 @@ class mod_bigbluebuttonbn_lib_testcase extends advanced_testcase {
         $event->timestart = time();
 
         return calendar_event::create($event);
-    }
-
-    public function test_bigbluebuttonbn_log() {
-        global $DB;
-        $this->resetAfterTest();
-        list($bbactivitycontext, $bbactivitycm, $bbactivity) = $this->create_instance();
-        bigbluebuttonbn_log($bbactivity, BIGBLUEBUTTONBN_LOG_EVENT_PLAYED);
-        $this->assertTrue($DB->record_exists('bigbluebuttonbn_logs', array('bigbluebuttonbnid' => $bbactivity->id)));
     }
 
     public function test_bigbluebuttonbn_extend_settings_navigation_admin() {
