+++ tests/recordings_test.php
@@ -26,7 +26,6 @@
 defined('MOODLE_INTERNAL') || die();
 global $CFG;
 
-require_once($CFG->dirroot . '/mod/bigbluebuttonbn/lib.php');
 require_once($CFG->dirroot . '/mod/bigbluebuttonbn/locallib.php');
 
 /**
